+++ projects/create-remult/package.json
@@ -42,7 +42,7 @@
     "prepublishOnly": "npm run build",
     "test": "vitest --run",
     "try": "npx create-remult remult-project1 -t next -s express -d postgres --crud --overwrite yes --admin -a auth.js",
-    "try-l": "npm run build && node ./dist/index.mjs remult-project-sveltekit -t angular -d json --crud --overwrite yes --admin -a auth.js"
+    "try-l": "npm run build && node ./dist/index.mjs remult-project-sveltekit -t angular -s express-vite -d json --crud --overwrite yes --admin -a auth.js"
   },
   "engines": {
     "node": "^18.0.0 || >=20.0.0"
